+++ code/src/cljs/sixsq/nuvla/ui/utils/map.cljs
@@ -34,6 +34,11 @@
                 :attribution "&copy; <a href=\"http://osm.org/copyright\">OpenStreetMap</a>"}]]])
 
 
+(defn normalize-lng
+  [lng]
+  (-> lng (mod 360) (+ 540) (mod 360) (- 180)))
+
+
 (defn convert-latlong-map
   [latlong]
   [(.-lng latlong) (.-lat latlong)])