+++ code/src/cljs/sixsq/nuvla/ui/edge/views.cljs
@@ -239,8 +239,8 @@
         uuid     (general-utils/id->uuid id)
         on-click #(dispatch [::history-events/navigate (str "edge/" uuid)])]
     [map/CircleMarker {:on-click on-click
-                       :center   location
-                       :color    (utils/status->color @status)
+                       :center   (map/longlat->latlong location)
+                       :color    (utils/map-status->color @status)
                        :opacity  0.5
                        :weight   2}
      [map/Tooltip (or name id)]]))
@@ -288,7 +288,8 @@
                     :cards [NuvlaboxCards]
                     :table [NuvlaboxTable]
                     :map [NuvlaboxMap])
-                  [Pagination]
+                  (when-not (= @view-type :map)
+                    [Pagination])
                   [AddModal]]
         children (case n
                    1 root