+++ code/src/cljs/sixsq/nuvla/ui/edge_detail/views.cljs
@@ -17,7 +17,8 @@
     [sixsq.nuvla.ui.utils.plot :as plot]
     [sixsq.nuvla.ui.utils.semantic-ui :as ui]
     [sixsq.nuvla.ui.utils.semantic-ui-extensions :as uix]
-    [sixsq.nuvla.ui.utils.time :as time]))
+    [sixsq.nuvla.ui.utils.time :as time]
+    [sixsq.nuvla.ui.utils.style :as style]))
 
 
 (def refresh-action-id :nuvlabox-get-nuvlabox)
@@ -159,30 +160,46 @@
 
 (defn LocationAccordion
   [{:keys [id location] :as nuvlabox}]
-  (let [zoom (atom 3)
-        tr   (subscribe [::i18n-subs/tr])]
+  (let [tr           (subscribe [::i18n-subs/tr])
+        zoom         (atom 3)
+        new-location (r/atom nil)]
     (fn [{:keys [id location] :as nuvlabox}]
-      (let [update-location-fn #(dispatch [::events/edit id (assoc nuvlabox :location %)])]
-        [uix/Accordion
+      (let [update-new-location #(reset! new-location %)
+            position            (some-> (or @new-location location) map/longlat->latlong)]
 
-         [:div
-          (if location (@tr [:map-drag-to-update-nb-location])
+        [uix/Accordion
+         [ui/Segment style/basic
+          (if position (@tr [:map-drag-to-update-nb-location])
                        (@tr [:map-click-to-set-nb-location]))
           [map/Map
-           {:style             {:height 400
-                                :width  "100%"
-                                :cursor (when-not location "pointer")}
-            :center            (or location map/sixsq-latlng)
+           {:style             {:height        400
+                                :width         "100%"
+                                :margin-bottom 10
+                                :cursor        (when-not location "pointer")}
+            :center            (or position map/sixsq-latlng)
             :zoom              @zoom
             :onViewportChanged #(reset! zoom (.-zoom %))
-            :on-click          (when-not location
-                                 (map/click-location update-location-fn))}
+            :on-click          (when-not position
+                                 (map/click-location update-new-location))}
            [map/DefaultLayers]
 
-           (when location
-             [map/Marker {:position    location
+           (when position
+             [map/Marker {:position    position
                           :draggable   true
-                          :on-drag-end (map/drag-end-location update-location-fn)}])]]
+                          :on-drag-end (map/drag-end-location update-new-location)}])]
+          [:div
+           [ui/Button {:primary  true
+                       :floated  "right"
+                       :on-click #(dispatch
+                                    [::events/edit id
+                                     (assoc nuvlabox
+                                       :location
+                                       (update @new-location 0 map/normalize-lng))
+                                     "NuvlaBox position updated successfully"])}
+            (@tr [:save])]
+           [ui/Button {:floated  "right"
+                       :on-click #(reset! new-location nil)}
+            (@tr [:cancel])]]]
          :default-open false
          :label "Location"
          :icon "map"]))))
@@ -265,7 +282,7 @@
 
         [ui/CardMeta (str (@tr [:created]) " " (-> created time/parse-iso8601 time/ago))]
 
-        [:p {:style {:float "right"}} state]
+        [:p {:align "right"} state]
 
         (when-not (str/blank? description)
           [ui/CardDescription {:style {:overflow "hidden" :max-height "100px"}} description])
@@ -310,6 +327,7 @@
            :default-active? @acl-open
            :on-change       #(do
                                (reset! acl-open true)
-                               (dispatch [::events/edit (:id @nuvlabox) (assoc @nuvlabox :acl %)]))}])
+                               (dispatch [::events/edit (:id @nuvlabox) (assoc @nuvlabox :acl %)
+                                          "NuvlaBox ACL updated successfully"]))}])
        [SummarySection]
        [StatusSection]])))