+++ code/src/cljs/sixsq/nuvla/ui/ocre/views.cljs
@@ -383,8 +383,8 @@
 (defn View
   []
   (dispatch [::cimi-events/set-selected-fields
-             ["code", "amount", "currency", "platform", "target-audience",
-              "state", "created", "distributor"]])
+             ["code", "amount", "currency", "supplier", "country-name",
+              "state", "created"]])
   (dispatch [::cimi-events/set-collection-name "voucher"])
   (dispatch [::cimi-events/set-filter nil])
   (dispatch [::cimi-events/set-aggregation nil])